Bug #16089: Fix vtkmaskPoints multi-process missing condition
In old verion, if proportionalMaximumNumberOfPoints is enabled with one
process, localMaxPts would be set to MaximumNumberOfPoints which obsolete
the upper line. After adding the process number checking condition, the if
statement in GetLocalSampleSize function is no longer needed. And we also
need to compare maximumNumberOfPoints with the total for garbage memory
safety reason.