Ignore nested lists following scalar data in openFOAM lists.
vtkOpenFOAMReader would fail to read files if a list of the form:

(scalar (scalar scalar) )

was found, as it only supported nested lists to appear before all
'loose' scalars. This patch allows such files to be read by discarding
the nested list data and printing a warning, instead of failing to
finish reading the file.

This is to work around an issue reported on the mailing list where such
a list was used in a field that was not needed for visualization.