fix cell coloring of degenerate polygons
gl_PrimitiveID seems ot not get incremented for
degenerate polygons consistently even if the
point IDs are different but the ocations are the same.
This topic skips rendering any polys that are
degenerate based on point locations. It regresses
back to OpenGL1 how it triangulates polygons to a
simple algorithm as opposed to using vtkPolygon