VTK Bug15270: Correct empty condition for vtkGlyph3DMapper
When working on an empty polydata, vtkGlyph3DMapper::GetBounds would
return the bounds of the glyph instead of vtkMath::UninitializedBounds.
This would be confusing because glyph's bounds has nothing to do with
the rest of the scene and discrupts
vtkRenderer::ResetCameraClippingRange. To fix this, a check condition is
added and a new test file is included.