ENH: New boolean operator filter
This is a rebased commit of Adam Updegrove's original MR. Also the
code and test have been indendented according to the new VTK
indentaion rules. And the doxygen has been converted to the new
doxygen rules. Other than those changes, the rest of the code is
Adam's.

We realize there is additonal work to be done on this filter to
improve its robustness, but the development team feels it is an
important contribution and will be easier to improve onve merged into
VTK proper.

Here are Adam's original comments regarding the MR.

This is a new Boolean that is possible following the new code recently
added to vtkIntersectionPolyDataFilter. @cory-quammen @dgobbi

I have given the new boolean the name vtkLoopBooleanPolyDataFilter as
it finds intersection loops and runs around these in an oriented
fashion in order to determine the sub-surfaces of the Boolean. I felt
this was better than vtkBooleanOperationPolyDataFilter2, especially
since there was a test already named that way. I'm open to changing
the name though.

Oh, and as a plus, I removed the vtkDataSetSurfaceFilter that I was
previously using, so I no longer need to bring in the Filters/Geometry
module. Makes the change a lot smoother and simpler.

Again, more details on this can be found here:
http://www.sciencedirect.com/science/article/pii/S0965997816300230