Add ForceOpaque and ForceTranslucent to vtkImageActor
The vtkImageActor checks for the presence of an alpha component to see if
translucent rendering should be used.  This can be problematic, because
even if there is an alpha component, the alpha component might be 255 for
all voxels and the actor might therefore be opaque.

A better option is to allow the user to declare their intent by calling
ForceOpaque or ForceTranslucent.

The base class, vtkImageSlice, only has the ForceTranslucent option
because it always predictably renders in the Opaque pass unless forced
to render in the Translucent pass.