Added algorithms to mask NaNs.
VTK does not natively handle NaNs. To enable a preliminary support
for this, I added a Python algorithm to create a ghost array
based on NaN values. This algorithm creates an array which masks
using HIDDENPOINT or HIDDENCELL values points or cells containing
NaN values. This array can then be used to either threshold or
simply as a ghost array.