Fix order of representations in a view.
vtkView keeps representations provided to `AddRepresentation` call in an
internal datastructure. If a representation was comprised of other
internal representations and it called vtkView::AddRepresentation() then
the order in which the representations would get stored in the internal
datastructure did not match the order in which `AddRepresentation` was
called. Not a big deal, except in cases where the internal
representation expects the outer representation to have been updated (as
was was the case with #16832).

Fixed vtkView::AddRepresentation() to maintain the order in internal
datastructure to match the order in which AddRepresentation is called.

Fixes #16832.