ENH: Add vtkTextProperty::UseTightBoundingBox to center a label to anchor.
This property allows the user to center the string to the anchor point.
By default, the bounding box for the string is computed
using the font metrics which includes ascents and descents.
As descents might not be present in the string, the label will not
be perfectly centered on the anchor point. Setting this property
on will compute the bounding box using the current string rather than
the font metric. This results in perfectly centered labels.
This does not work well for temporal data if the string changes as
it results in text that moves around.