some tweaks and fixes for VR
A few tweaks and fixes as detailed below.

Remove the use of a physical scale factor in
computing the camera transforms. Instead use the
distance from the focal point to the camera as a
metric. That distance will be mapped into one
meter in the physical space. This fits the model
of a camera dolly.

This has the advantage of removing the scale factor that
was being added into the World to View matrix. We generally
assume there is not a scale in the world to view matrix and
that was causing some issues with some shaders (such as imposters)

Fix an issue where the scale was being added to the NormalMatrix
The normal matrix should never have a scale.

Adjust the scaling interaction to scale about a meter above the
floor. This is more intuitive than scaling from the floor as an
origin.

Fix a bug in the Oculus code when a buffer submission failed with
a resubmit success error code.