Add vtkImageConnectivityFilter
This is the vtkImageData equivalent of vtkPolyDataConnectivityFilter.
It performs connectivity checks on points with values within a specific
scalar range (by default, it checks all points with values > 0).  Its
output is a label image, where each connected region is assigned an
integer label value.  Unlabeled points are assigned a value of zero.