Added scale factor for distance representation
Added ability to set the scale factor from VTK world coordinates. The
tick marks and label will be defined in terms of the scaled space. For
example, if the VTK world coordinates are assumed to be in inches, but
the desired distance units should be defined in terms of centimeters,
the scale factor should be set to 0.3937. The tick marks will then be
spaced in terms of centimeters, and the label will contain the
measurement in centimeters.