Fixing ExodusIIWriter segfault with only float arrays.
If there are no double arrays in the input, the attributesD
vector is not resized to contain pointers for all blocks, even
though a pointer is always written to this vector when processing
each block (it may be null). This resulted in a segmentation fault.
The problem is fixed by always setting the size of the attributesD
vector of double array pointers to the number of blocks.