Call glFlush() after initializing a new context.
There were some odd behaviors on ParaView when using pvserver with
--use-offscreen-rendering (see #16943). The initial render on a new
context would result in an empty scene. This behavior started when
calls to glFlush() in vtkRenderWindow::Frame() were removed for
performance reasons.

This patch fixes the rendering issue, and since it only happens
for new contexts, won't hurt performance.