Modify behavior of average intensity projected volume rendering
The behavior is now consistent with additive projection where the
scalar values are multiplied by opacity calculated from the opacity
transfer function and then added. The additional step here is to
divide the sum by the number of samples taken through the volume.

As is the case with the additive intensity projection, the final
image will always be grayscale i.e. the aggregated values are not
passed through the color transfer function. This is because the
resultant value is a derived value and not a real data value along
the sampling ray.