Improve bounds when orienting by an array
To handle arbitrary glyph orientations, reflect the bounds of the
glyph across the origin in each dimension. The union of the reflected
bounds and original bounds will then bound any orientation of the
glyph. This union is automatically centered at the origin, so there
is no need to translate it by the center of the bounding box.