Fixed stochastic jittering and added API to customize noise settings.
The mapper uses vtkPerlinNoise by default with a texture size and frequency
equivalent to the window size.  It is possible to fully customize the noise
generator and texture size.  UseJittering changed to be off by default so some
of the baselines were updated.  Added tests for both default and customized noise
ray jittering.  Disabled TestGPURayCastMapperSampleDistance in legacy OpenGL.