combine fbo and fbo2 into one class and clean it up
having two VBO classes was confusing. This topic
moves the functionality of FBO2 into FBO while
keeping some of the original FBO functionality and
cleaning it up a bit.

All OpenGL2 code in VTK using FBOs is updated in
this topic to use the merged FBO class. Mainly this
impacts classes that were using the FBO class.

The old FBO methods and code are converted but
commented out in this commit.