Fix issue with coincident geometry + MSAA. See #16899.
101f9eeb6e8dd1834f39f4c5180cfc04675b8bca changed the polydata
fragment shaders to always write out gl_FragDepth explicitly to help
with the depth peeling abstractions. However, this causes issues when
multisampling, as the multisampled depths are not handled correctly when
gl_FragDepth is not the default value. [1] has some more detailed info.

After this patch, gl_FragDepth is only set when it is needed.

[1] http://www.gamedev.net/topic/544920-gl_fragdepth-and-anti-aliasing/#entry4514289