add in part motion, and some cleanup
We now look for a list of parts, that is blocks that make up one
logical component, and then at each time step a translation value
for all the blocks in the part.

Also, use hdf element detection to try and silence the CanRead
test warning from ParaView which exercises that bit of code more.