Picking classes can pick a vtkCompositeDataSet
The picker classes now successfully pick a vtkActor that uses a mapper
taking a vtkCompositeData as input.  Before this change picking would
fail.

The composite dataset and the flat block index of the block that is
picked are stored as pick result in vtkPicker. These will have
non-default values (stored by a new MarkPicked routine in vtkPicker)
only when the picker successfully picks a composite dataset.

vtkPicker will return the nearest block's dataset. The pick point is the
nearest block's center instead of the composite dataset's center.

In vtkCellPicker and vtkPointPicker a new protected method is does the
intersection of the ray with a vtkDataSet. The method for intersecting
an actor calls this routine, and does so iteratively for all intersected
blocks if the mapper input is composite.

pickCompositeData.py tests prop, cell and point picking.