Fixed a bug where GetScreenSize on Cocoa was broken
On Cocoa, the GetScreenSize() method was basically nonsense,
and would very likely crash on multi monitor systems.

It was indexing the [NSScreen screens] array with the currentVirtualScreen
value, but I can find no docs to support that this is at all reasonable.

Instead, get the NSScreen directly from the NSWindow.

Also synced a comment between OGL1 and OGL2 code.