Removed unused local variable from nifti writer.
The vtkNIFTIImageWriter had a local variable "swapBytes" that was always
equal to zero.  The original intent was to allow the endianness of the
file to be chosen, but this was never implemented.  The writer always uses
the native architecture endianness for the file, which is standard for
NIFTI since the header specifies the endianness.