many fixes and new features
Add floor option to show a floor

completely rework the matrix and coordinate code
to be cleaner and use standard VTK conventions (right handed)

add support for a dashboard overlay with options to subclass
and support for saving/loading camera poses and adjusting
motion factors and scale factors.

Improve the resetcameraclippingrange to work better for VR

change how eye position is handled so that VolumeRendering
will work properly

Handle the case where poeple make the cubes invisible which
results in a 308 error on loading the model.

Add support for the user to specify the mapping of physical
coordinates into world coordinate. Allows for different
view up vectors other than the old code which was hard
coded to Y.

Add middle mouse support mapped to the grip button.

Add some support for a hardware clipping plane.