Added support for reader/writing composite field data.
The XML readers and writers did not previously support reading
and writing of field data in composite data. Added support.
This is a first pass at this. Only field data of the root node
is supported and it is always written inline - that composite
files do not support appended data. This should be used for
meta-data mainly not large field data.

Also changed the composite pipeline to copy field data for
composite data when the filter is not composite aware. Previously
the output was initialized and field data was lost.