Remove duplicate "OutputSpacing" member.
The class vtkImageResample declared a member called OutputSpacing, even
though that member was already present in the superclass vtkImageReslice.
In order to ensure that the SetOutputSpacing() method (in vtkImageReslice)
can be used with vtkImageResample, it is necessary that vtkImageResample
does not redeclare this member.

Note that vtkImageResample uses a spacing of [0,0,0] as a signal that the
spacing is "undefined", whereas vtkImageReslice does not (it uses a default
spacing of [1,1,1] and a flag called ComputeOutputSpacing to indicate
whether the filter should use these values or compute new ones).  This
doesn't cause a bug since the OutputSpacing member only used in
RequestInformation, and each class has its own RequestInformation
implementation.