16849: Wrap classes that have no copy constructor.
Previously, if a class in VTK had no copy constructor it would not be
wrapped (unless it was derived from vtkObjectBase, which always has the
copy constructor disabled).

Examples of classes that are now wrapped are vtkFloatingPointException
(which only has static methods and is never actually instantiated) and
RAII classes like vtkSimpleMutexLock that lack copy constructors for
obvious reasons.