reduce segfaults on shader compile failure
When a shader program fails to compile often the
application would crash because code assumed that the
program was created. The shader compilation error messages
often were hard to obtain as the segault kills the output
window.

This topic catches many places where the program was assumed
to exists and makes them not crash. It was developed by
modifying the shader cache to always fail when building a
shader and then running the VTK tests to see if anything
crashed. Lots did. With this topic at least

ctest -R OpenGL2
ctest -R RenderingCore
ctest -R Chart

all work without segaulting even when the shader fails
to compile. This makes catching the shader compilation
error messages lots easier and of course less chance of
an application segfaulting as opposed to gracefully
reporting an error.