Optimization for vtkMultiBlockVolumeMapper.
vtkMultiBlockVolumeMapper now uses various internal mappers to render
(one per block), hence avoiding the need to reload data every time a
different block is rendered. The mapper checks at loading time whether
all of the blocks would fit in GPU memory and if so, creates a mapper per
block. If this is not the case, the mapper falls back to the single
internal mapper approach.