Fix translucent geometry outside of volume bounds when volume peeling.
The front peels in the dual depth peeling algorithm assumed that there
would be a shader invocation for each pixel that is affected by the
peeled geometry. This is because the front peels are blended in the
shader and always written to a double buffered accumulation texture.
If there are no more fragments to be processed for a particular pixel,
we'd still execute a shader on that pixel, which would simply pass the
color from the input on to the output.

However, now that we're splitting the rendering up into two
passes-per-peel (one for translucent geometry, one for volumes), we
can no longer rely on this in-shader pass-through, since translucent
geometry may lie outside the volume boundaries. Such geometry will not
be propagated to the output texture, resulting in rendering artifacts.

This patch addresses this problem by initializing the front destination
buffer to match the front source buffer before each render pass. Now
pixels that aren't shaded by a particular pass will still contain the
colors written from the previous pass.

Since a clear is going to be faster than the fullscreen render that is
used to copy the texture, the copy only happens when volume peeling is
enabled, and falls back to the fast glClear when we can get away with it.