Track whether glewInit succeeded better
Added a new member variable which records success or failure, use that
variable in the SupportsOpenGL function (if glewInit fails then it is
likely nothing will work). Added an #else if the block for when GLEW is
not used, and marked it as true so long as OpenGLInitContext has been
called. In my local testing this worked as expected when triggering
failure conditions.