Fix QVTKInteractorAdaptor mouse wheel events
There were two problems that needed to be fixed for the wheel events.
The first problem had to do with changing direction: since the
AccumulatedDelta kept the residual after the threshold was passed,
the first wheel click in the opposite direction was sometimes ignored.
Not keeping any residual fixes this, though it makes scrolling on a Mac
trackpad a bit less smooth.  The second problem is that the delta()
method is deprecated for Qt5, and angleDelta() should be used instead.