adapt findospray for ospray version 1.2
starting in ospray 1.2, embree is no longer packaged and you must
build it separately. This change fixes the "build vtk with
ospray's build directory" option. The "build vtk with ospray's
install directory" option appears to work as is, at least in the
case I've tested where both ospray and embree are installed
in the same place.