Adjust the cylinder widget's radius changing interaction
The cylinder widget already computes the delta in world coordinates so
there is no need to scale by the length of the dataset again.  Doing so
was creating some very difficult to adjust interaction with very large
and very small datasets.  Dividing by a constant seems to keep the
interaction appropriately scaled across many dataset sizes.