Add vtkOpenVRRay to vtkOpenVRModel
This commit adds a way to draw a ray for a particular model. This will
be useful when picking is performed along a ray using the
vtkOpenVRPropPicker introduced in 653fee20c0eab618074868666874e8a3a7c0c1c0 .
The length of the ray can be set from the model and corresponds to a
distance in meters. To do so, you can access the different models using
the accessor introduced in 0ea021ed34c7241fa9ac5f499d9cc3f858f422ce .