Improved CopyStructuredData() to handle more cases.
CopyStructuredData() only supported the use case of copying
from a larger structured dataset to a smaller one. This was
added to handle output initialization of imaging algorithms.
Extended the method to support cases where the two structured
data are of arbitrary extents as long as they overlap. This
is useful in various situations. For example, when writing
ghost cell data exchange code, one has to receive a small
structured data containing the ghost data and then copy those
values to the larger structured data. For this use case, I
also added SetupForCopy() which sets up the mapping between
input and output arrays without having to call CopyAllocate()
which wipes out any existing data.