Conserve color attachments in DDP.
The dual depth peeling algorithm used to bind all working
textures to the framebuffer and only activate them as needed.
This was causing issues on some OpenGL ES implementations
which limit the number of color attachments to 4, while DDP
was trying to bind 6 of them.

At most, only 3 color attachments are actually used
simultaneously, and this patch refactors the code to only
bind the color attachment textures as needed.

Benchmarking shows that this adds a negligible 0.5% to the
render time for our DP benchmarks.