add a framebuffer pass to support rendering into a fo
This way you can render into a Frambuffer Object and then
at the last step it does a blit to the outer framebuffer.
Typically the screen/window but it could be another FO.

Also add a FramebufferDepthPeelingPass that is desgined to
work within a framebuffer pass (or similar). This is the
only depth peeling that will work on OpenGL ES3 right now.