When wrapping, use template args for scoping constants.
When the wrappers use a constant that is defined within a class or
namespace, the constant must be scoped, e.g. mynamespace::MY_VALUE.
This was working fine, except for when the scope was a templated
class, e.g. myclass<class T1, int N1>, because the wrappers were
ignoring the template args.  Now, the wrappers include the template
args in the scope, e.g. myclass<T1,N1>::MY_VALUE.