Fix iOS build with CMake 3.8
The iOS build (i.e. specifying -DVTK_IOS_BUILD:BOOL=ON) failed with an
error that the vtkEncodeString utility is not found. The failure was
first introduced in commit 894acce.

This commit changes the logic to import VTKCompileTools when
CMAKE_CROSSCOMPILING_EMULATOR is not defined even if CMake fully
supports it.