restore ability to follow GridControllers
Now we only do the heavy read when required but we still
get the data when we read the top level file.

Todo: xdmfgridcontroller still assumes that the destination is
an unstructured grid. That needs fixing in libxdmf.
Todo: The parallel writer only considers RECTGRID and UNSTRGRID
when doing parallel writes.