Enable vtkPlaneCutter without SphereTree
This commit enables to use vtkPlaneCutter
without using a vtkSphereTree. When slicing,
if the sphere tree is not available, an array of
in/out value will be computed for each points,
wich allows an easy selection of cells. This mode is not as fast
as the sphere tree but does not require to build
any structure with the first slice.
This also corrects a bug with cellData construction
in SMP and updates the tests.