Allowing gradient and CD2PD filters to work on specific cell dimensions
The cell dimension for unstructured grids and polydata can vary. These changes allow
specifying the cell dimensions to use when performing the computations. The default
behavior should be exactly the same as before. Also getting rid of the GetVoidPointer()
calls on the input array. Using them on the output array should be fine though since
the arrays are explicitly created with the AOS layout. Also, simplified the
Q-criterion computation. This modification also makes the output arrays be either
vtkFloatArrays or vtkDoubleArrays.

Addresses paraview/paraview#16931.