COMP: Mac issues with odd-sized baseline images
NSWindow's behaviour round sizes *up* to even integral sizes.

This seems to be related to the limitation for Macs with retina
display - they can only create windows with width and height with even
number of pixels.

If you try to create a window with height or width
with odd number of pixels you end up with height or width with even
number of pixels.