Replace many NULL and 0 occurances with nullptr
This topic is the result of running clang-tidy to
modernize our usage of NULL and 0 to nullptr. It also
includes some manual and semi manual changes where
clang-tidy would not be expected to work (such as in
comments, or classes not compiled on this build)

There are definitely many comments and occurances that
this topic misses that we will need to fix over time.