fix bug with single arrays that should auto shift scale
There was a bug where is auto shift scale was turned on
when uploading a single array (as opposed to appending) the
shift scale factor was not computed prior to testing for
the fast path. This caused some arrays to be uploaded in their
non shift scaled form when shift scale should have been on.

Fixing this exposed bugs in RectangularBUttonSource
and ExtractSurface producing uninitialized points

ExtractSurface is being worked on by Will to fix the issue.