Added support for XYSampling with depth-peeling on.
When an internal FBO (reduced in size) is used by the volume mapper, fragment
coordinates need to be scaled to the original size in order to fetch correct
texture values. Additional targets may be required when rendering the final
quad to output depending on the current render stage in the RenderPass (e.g.
Peeling needs 2 to render front and back buffers while AlphaBlending only
one).