Rebuild representations when the RenderWindow is updated
The size of the handles in the vtkImplicitPlaneRepresentation and
vtkImplicitCylinderRepresentation were not being updated when the
RenderWindow changed size. This could make the handles smaller or
larger than they should be in some cases.  Rebuilding the
representation when the RenderWindow changes ensures the handles are
the correct size in such cases.