Extended vtkOpenGLRenderPass API to track the number of DrawBuffers activated.
vtkOpenGLRenderPass provides the number of activated DrawBuffers in case the
mapper needs to change the targets during rendering. GPUVolumeRCMapper for
instance, in some cases, renders to an internal FBO in order to reduce the
number of casted rays.

Conflicts:
	Rendering/OpenGL2/vtkDualDepthPeelingPass.cxx

Included changes in the base class (Rendering/OpenGL2/vtkDepthPeelingPass.*)
required for this patch in DualDepthPeeling.