Reduce VTK module output during CMake configuration
The reporting of enabled VTK modules adds significant output that is
not typically useful during the CMake configuration step and makes it
more difficult to find new configuration warnings.

Write the enabled VTK modules to a file in
${VTK_BINARY_DIR}/CMakeFiles/VTKModules.log instead of printing them
to stdout.

disables the module dependency output by default. It can be
re-enabled by setting the value of the VTK_REPORT_ENABLED_MODULES option
to ON, e.g., -DVTK_PRINT_MODULES:BOOL=ON.