Lights: clear transform when switching types. Fix typos.
A camera light uses the transform to match the camera's position,
then applies an offset. The transform matrix must be cleared when
switching to head/scene light for them to position correctly.
Always clear the transform matrix when switching types.

Fix a duplicate include and make coneAngle consistent between
ospray and opengl.