update handling of scalar material mode
Previously scalar material mode when used
with scalars would ignore the ambient and diffuse
material intensities and when mapping one of them
set that one to 1.0 while leaving th eother alone.
This caused lots of odd situations where objects would be
over illuminated etc or where changing the diffuse or
ambient seemed to have o impact.

With this change the diffuse and ambient intensities
are always used. The sclaar simply controls the color
not the intensity.