deprecate scalar material mode
The old scalar material mode was largely unused or used in
an inappropriate manner and yet adds complexity and cost
to the code. Thsi topic removes it and replaces it with
an implementation that is similar to
VTK_MATERIALMODE_AMBIENT_AND_DIFFUSE so that scalar colors
are used for ambient and diffuse colors but the ambient
and diffuse intensities are still respected.