vtkPolygon: Add unbiased ear cut and bounded triangulate methods.
vtkPolygon's EarCutTriangulation() method uses a priority queue to
determine the order in which polygon vertices are clipped. This
results in a reproducible triangulation that is orientation-free, but
only allows for a single triangulation for a polygon. Since
ear-cutting is a heuristic, there is no guarantee that the result from
EarCutTriangulation() is the best (or even acceptable).

UnbiasedEarCutTriangulation() is an ear cutting method that removes
the use of a priority queue, and instead allows the caller to seed the
method's start point with a value.

With the ability to triangulate a polygon in multiple ways,
BoundedTriangulate() accepts a tolerance value that restricts the
ratio of the area of the smallest triangle to the area of the
polygon. The method returns the first triangulation that satisfies
this criterion.