fix for GetOrientation on scaled matrices
The old code ignored scale when evaluating if an axis should
be ignored (1.0e-4) But you need to consider scale as it
impacts the dot product used in the comparison. 1,0e-5 may be the
largest axis of the matrix, not something to ignore. This
fixes that issue. I am not sure that heterogeneous scales
are correctly handled by the original code or this code, but
at least homogeneous scales should work better with this.