Predictable output for vtkTensorGlyph on non-symmetric tensors
Previous version implicitly assumed that the tensor was symmetric
when using eigenvalue decomposition. General tensors have
non-real eigenvalues and eigenvectors.
It did not crash on general input because vtkMath::Jacobi expects
real symmetric matrix, and only use the upper right triangular
part of the matrix. So computations were as if the lower
triangular part of the tensor was symmetric of its the upper
triangular part.
It make more sense to apply the eigenvalue decomposition to the
symmetrical part of the tensor.