Introduce custom transfer function range type to volume mapper
This change introduces a new variable called TFRangeType for each of the
three transfer functions of the volume mapper. When set to SCALAR, the
existing functionality of the mapper is preserved, where the function is
scaled over the scalar range. When set to NATIVE, the mapper scales the
function over the range set on the transfer function i.e. min and max of
all the function points. Any scalar values outside this range would be
truncated to the min and max of native range.