Fix vtkCellPicker use of locator plus clipping planes
When a cell locator was used with vtkCellPicker, an incorrect
result was returned if the actor's mapper had clipping planes.
This bug was introduced when vtkCellPicker was restructured in
January, 2017 (see 3eeee1ad).

A second clipping plane has been added to the PickerWithLocator
test so that it will detect similar bugs.

Fixes #17143