Ensure that all processes in subgroup invoke collectives
In a distributed dataset where some processes have datasets with no
points or cells, not all processes would invoke the same collective
communication operations in
vtkPKdTree::CreateGlobalDataArrayBounds(). This would lead to a
deadlock in pvserver where processes waited for messages that never
came.

This is solved by distributing the maximum number of cell and point
data arrays to all processes so that they all participate in the
collective operations.