Fixed a bug in the Moore supercursor
The goal of this commit was to fix a bug that occurred only in
the (f=2;d=3) case ("binary 3D"), and that had been caused
by a typo in the MooreChildCursorToChildTable and
 MooreChildCursorToParentCursorTable tables: specifically,
a "23" had been written instread of "32".

In addition, two non-regression tests, as well as their baseline
images, were added in order to validate this fix:
(i) a geometry case which does not exercise the offending part of
the code but that illustrates the case built to do it; and
(ii) a contouring case that does exercise that part of the code,
for it relies on the "32" section of the Moore supercursor tables.